+++ CHANGELOG.md
@@ -221,3 +221,11 @@
 
 - Moved some nullable things from the creative tab's builder to the BuildCreativeModeTabContentsEvent, where hopefully they won't be null, and fixed an NPE if it is
 - Changed Gene.PotionDetails.CODEC to use `show_icon` instead of `showIcon`
+
+# 1.4.0
+
+- Added the Gene Checker, which lets you see the Genes of either yourself or the entity you're looking at
+- Changed some tags:
+    - `#geneticsresequenced:syringe` is now `#geneticsresequenced:syringes`
+    - `#geneticsresequenced:magnet_blacklist` is now `#geneticsresequenced:item_magnet_gene_blacklist`
+- Removed the tag `#geneticsresequenced:wooly`, it just uses `#c:tools/shear` now